+++ back-end-projects/OntSynHandler/src/main/java/com/github/ontio/model/dao/Oep5.java
@@ -5,6 +5,8 @@ import lombok.Builder;
 import javax.persistence.Column;
 import javax.persistence.Id;
 import javax.persistence.Table;
+
+import java.math.BigDecimal;
 import java.util.Date;
 
 @Table(name = "tbl_oep5")
@@ -25,7 +27,7 @@ public class Oep5 {
      * OEP5代币总量
      */
     @Column(name = "total_supply")
-    private Long totalSupply;
+    private BigDecimal totalSupply;
 
     /**
      * OEP5代币符号
@@ -51,7 +53,7 @@ public class Oep5 {
     private Date updateTime;
 
     @Builder
-    public Oep5(String contractHash, String name, Long totalSupply, String symbol, Date createTime, Boolean auditFlag, Date updateTime) {
+    public Oep5(String contractHash, String name, BigDecimal totalSupply, String symbol, Date createTime, Boolean auditFlag, Date updateTime) {
         this.contractHash = contractHash;
         this.name = name;
         this.totalSupply = totalSupply;
@@ -102,7 +104,7 @@ public class Oep5 {
      *
      * @return total_supply - OEP5代币总量
      */
-    public Long getTotalSupply() {
+    public BigDecimal getTotalSupply() {
         return totalSupply;
     }
 
@@ -111,7 +113,7 @@ public class Oep5 {
      *
      * @param totalSupply OEP5代币总量
      */
-    public void setTotalSupply(Long totalSupply) {
+    public void setTotalSupply(BigDecimal totalSupply) {
         this.totalSupply = totalSupply;
     }
